refactor/total-portfolio-balance-composable
